fix: correct overflow bugs for long sequences

+++ csrc/causal_conv1d_bwd.cu
@@ -362,8 +362,9 @@ void causal_conv1d_channellast_bwd_kernel(ConvParamsBwd params) {
         + (chunk_l_id * kChunkSizeL + l_idx) * params.dx_l_stride + chunk_c_id * kChunkSizeC + c_idx * kNElts;
     float *dweight = reinterpret_cast<float *>(params.dweight_ptr)
         + chunk_c_id * kChunkSizeC * params.dweight_c_stride;
+    // batch_id and params.seqlen are both int, so widen before multiplying.
     int *seq_idx = !kHasSeqIdx ? nullptr : reinterpret_cast<int *>(params.seq_idx_ptr)
-        + batch_id * params.seqlen + chunk_l_id * kChunkSizeL;
+        + static_cast<int64_t>(batch_id) * params.seqlen + chunk_l_id * kChunkSizeL;
     input_t *initial_states = params.initial_states_ptr == nullptr || chunk_l_id > 0 ? nullptr
         : reinterpret_cast<input_t *>(params.initial_states_ptr) + batch_id * params.initial_states_batch_stride + l_idx * params.initial_states_l_stride + chunk_c_id * kChunkSizeC + c_idx * kNElts;
     input_t *dinitial_states = params.dinitial_states_ptr == nullptr || chunk_l_id > 0 ? nullptr
@@ -510,8 +511,9 @@ void causal_conv1d_channellast_bwd_kernel(ConvParamsBwd params) {
         if (col_idx == 0 && chunk_c_id * kChunkSizeC + row_idx < params.dim) {
             if constexpr (kDeterministic) {
                 float *dbias_ws = reinterpret_cast<float *>(params.dbias_workspace_ptr);
+                // chunk_l_id and params.dim are both int, so widen before multiplying.
                 dbias_ws[batch_id * params.dbias_workspace_batch_stride
-                       + chunk_l_id * params.dim
+                       + static_cast<int64_t>(chunk_l_id) * params.dim
                        + (chunk_c_id * kChunkSizeC + row_idx)] = dbias_val;
             } else {
                 atomicAdd(&reinterpret_cast<float *>(params.dbias_ptr)[chunk_c_id * kChunkSizeC + row_idx], dbias_val);
